Fix integer narrowing on x86 (backport)

+++ src/base/utils/fs.cpp
@@ -354,7 +354,7 @@ bool Utils::Fs::isNetworkFileSystem(const QString &path)
     // Magic number references:
     // 1. /usr/include/linux/magic.h
     // 2. https://github.com/coreutils/coreutils/blob/master/src/stat.c
-    switch (buf.f_type) {
+    switch (static_cast<unsigned int>(buf.f_type)) {
     case 0xFF534D42:  // CIFS_MAGIC_NUMBER
     case 0x6969:  // NFS_SUPER_MAGIC
     case 0x517B:  // SMB_SUPER_MAGIC
